+++ R/coxphmulti.R
@@ -16,7 +16,7 @@
 #'   outputs. Output is of class \code{coxphlist}.
 #'
 #' @seealso \code{\link{fit2df}, \link{finalfit_merge}}
-#' @family \code{finalfit} model wrappers
+#' @family finalfit model wrappers
 #' @export
 #'
 #' @examples