+++ R/crrmulti.R
@@ -18,7 +18,7 @@
 #'   class \code{crr}.
 #'
 #' @seealso \code{\link{fit2df}, \link{finalfit_merge}}
-#' @family \code{finalfit} model wrappers
+#' @family finalfit model wrappers
 #' @export
 #'
 #' @examples