+++ R/crruni.R
@@ -18,7 +18,7 @@
 #'   \code{crrlist}.
 #'
 #' @seealso \code{\link{fit2df}, \link{finalfit_merge}}
-#' @family \code{finalfit} model wrappers
+#' @family finalfit model wrappers
 #' @export
 #'
 #' @examples