+++ R/glmmixed.R
@@ -20,7 +20,7 @@
 #'   Output is of class \code{glmerMod}.
 #'
 #' @seealso \code{\link{fit2df}, \link{finalfit_merge}}
-#' @family \code{finalfit} model wrappers
+#' @family finalfit model wrappers
 #' @export
 #'
 #' @examples