+++ R/glmmulti.R
@@ -22,7 +22,7 @@
 #'   outputs. Output is of class \code{glmlist}.
 #'
 #' @seealso \code{\link{fit2df}, \link{finalfit_merge}}
-#' @family \code{finalfit} model wrappers
+#' @family finalfit model wrappers
 #' @export
 #'
 #' @examples