+++ R/glmmultiboot.R
@@ -20,7 +20,7 @@
 #'   bootstrapped confidence intervals. Output is of class \code{glmboot}.
 #'
 #' @seealso \code{\link{fit2df}, \link{finalfit_merge}}
-#' @family \code{finalfit} model wrappers
+#' @family finalfit model wrappers
 #' @export
 #'
 #' @examples