+++ R/glmuni.R
@@ -16,7 +16,7 @@
 #'   Output is of class \code{glmlist}.
 #'
 #' @seealso \code{\link{fit2df}, \link{finalfit_merge}}
-#' @family \code{finalfit} model wrappers
+#' @family finalfit model wrappers
 #' @export
 #'
 #' @examples